Laptop 9/26
I was able to get the wave counter and countdown working, but I'm not sure how to hide them when the pause screen is up and there is a strange issue that occurs when all the waves have been finished. Also, I still need to figure out how to change the enemy drops and why the player's health starts at 1 when starting the level.